requirements.txt
fastapi
uvicorn[standard]
transformers
pydantic_settings
python-multipart
pytesseract
PyMuPDF
sentence_transformers
qdrant-client
fastembed